# crates/clawft-services/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WeftOS gateway API",
    "version": "0.8.1",
    "description": "REST, WebSocket and MCP surface of `weft gateway`. Every route except GET /api/health needs a bearer token issued by the kernel daemon (`weft token issue`); there is no mint route. Request and response bodies are mostly untyped here; see docs/ui/api-reference.md. Hand-maintained: tests/openapi.rs fails when a served route is missing or stale."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/agents": {
      "get": {
        "summary": "List agents",
        "operationId": "get_api_agents",
        "tags": [
          "handlers"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        }
      }
    },
...
    "/api/agents/{name}": {
      "get": {
        "summary": "Get agent",
        "operationId": "get_api_agents_name_",
        "tags": [
          "handlers"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "summary": "Delete agent by pid",
        "operationId": "delete_api_agents_name_",
        "tags": [
          "handlers"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/agents/{name}/start": {
      "post": {
        "summary": "Start agent",
        "operationId": "post_api_agents_name_start",
        "tags": [
          "handlers"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/agents/{name}/stop": {
      "post": {
        "summary": "Stop agent",
        "operationId": "post_api_agents_name_stop",
        "tags": [
          "handlers"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/auth/revoke": {
      "post": {
        "summary": "Revoke the caller's own token",
        "operationId": "post_api_auth_revoke",
        "tags": [
          "handlers"
        ],
        "responses": {
          "204": {
            "description": "Revoked"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        },
        "description": "Forwards the authenticating token's own id to the daemon. Cannot name another token."
      }
    },
...
    "/api/sessions": {
      "post": {
        "summary": "Create session",
        "operationId": "post_api_sessions",
        "tags": [
          "chat"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        }
      },
      "get": {
        "summary": "List sessions",
        "operationId": "get_api_sessions",
        "tags": [
          "handlers"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        }
      }
    },
    "/api/sessions/{key}": {
      "get": {
        "summary": "Get session",
        "operationId": "get_api_sessions_key_",
        "tags": [
          "handlers"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "summary": "Delete session",
        "operationId": "delete_api_sessions_key_",
        "tags": [
          "handlers"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/tools": {
      "get": {
        "summary": "List tools",
        "operationId": "get_api_tools",
        "tags": [
          "handlers"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        }
      }
    },
    "/api/tools/{name}/schema": {
      "get": {
        "summary": "Get tool schema",
        "operationId": "get_api_tools_name_schema",
        "tags": [
          "handlers"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/openapi.json": {
      "get": {
        "summary": "This document",
        "operationId": "get_api_openapi_json",
        "tags": [
          "openapi"
        ],
        "responses": {
          "200": {
            "description": "OpenAPI 3.1 document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Daemon-issued owner token (wft_...). Project tokens are refused."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing, unknown, expired, revoked or non-owner token",
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "DaemonUnavailable": {
        "description": "The daemon that owns tokens is unreachable",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "remedy": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  }
}